Move PointInCell related tests from experiemental to analysis
